Enable different logger implementations and optional logger
