+++ torch_em/trainer/default_trainer.py
@@ -5,11 +5,9 @@ import warnings
 import numpy as np
 import torch
 import torch.cuda.amp as amp
-from torchvision.utils import make_grid
-
 from tqdm import tqdm
 
-from ..util import ensure_tensor
+from .tensorboard_logger import TensorboardLogger
 
 
 class DefaultTrainer:
@@ -28,7 +26,8 @@ class DefaultTrainer:
         lr_scheduler=None,
         log_image_interval=100,
         mixed_precision=True,
-        early_stopping=None
+        early_stopping=None,
+        logger=TensorboardLogger
     ):
         self.name = name
         self.train_loader = train_loader
@@ -51,63 +50,8 @@ class DefaultTrainer:
         self.scaler = amp.GradScaler() if self.mixed_precision else None
         self.checkpoint_folder = os.path.join('./checkpoints', self.name)
 
-    def log_images(self, x, y, prediction, name, gradients=None):
-        step = self._iteration
-
-        def _normalize(im):
-            im = ensure_tensor(im, dtype=torch.float32)
-            im -= im.min()
-            im /= im.max()
-            return im
-
-        selection = np.s_[0] if x.ndim == 4 else np.s_[0, :, x.shape[2] // 2]
-
-        image = _normalize(x[selection].cpu())
-        self.tb.add_image(tag=f'{name}/input',
-                          img_tensor=image,
-                          global_step=step)
-
-        target_image = _normalize(y[selection].cpu())
-        pred_image = _normalize(prediction[selection].detach().cpu())
-
-        n_channels = target_image.shape[0]
-        if n_channels == 1:
-            nrow = 8
-            images = [image, target_image, pred_image]
-        else:
-            nrow = n_channels
-            images = nrow * [image]
-            images += [channel.unsqueeze(0) for channel in target_image]
-            images += [channel.unsqueeze(0) for channel in pred_image]
-
-        im_name = f'{name}/raw_targets_predictions'
-        if gradients is not None:
-            im_name += '_gradients'
-            grad_image = _normalize(gradients[selection].cpu())
-            if n_channels == 1:
-                images.append(grad_image)
-            else:
-                images += [channel.unsqueeze(0) for channel in grad_image]
-
-        im = make_grid(images, nrow=nrow, padding=4)
-        self.tb.add_image(tag=im_name,
-                          img_tensor=im,
-                          global_step=step)
-
-    def log_tensorboard_train(self, loss, lr, x, y, prediction, log_gradients=False):
-        step = self._iteration
-        self.tb.add_scalar(tag='train/loss', scalar_value=loss, global_step=step)
-        self.tb.add_scalar(tag='train/learning_rate', scalar_value=lr, global_step=step)
-        if step % self.log_image_interval == 0:
-            gradients = prediction.grad if log_gradients else None
-            self.log_images(x, y, prediction, 'train',
-                            gradients=gradients)
-
-    def log_tensorboard_validation(self, metric, loss, x, y, prediction):
-        step = self._iteration
-        self.tb.add_scalar(tag='validation/loss', scalar_value=loss, global_step=step)
-        self.tb.add_scalar(tag='validation/metric', scalar_value=metric, global_step=step)
-        self.log_images(x, y, prediction, 'validation')
+        log_dir = os.path.join('./logs', self.name)
+        self.logger = None if logger is None else logger(log_dir)
 
     def _initialize(self, iterations, load_from_checkpoint):
         assert self.train_loader is not None
@@ -130,8 +74,6 @@ class DefaultTrainer:
 
         os.makedirs(self.checkpoint_folder, exist_ok=True)
         os.makedirs('./logs', exist_ok=True)
-
-        self.tb = torch.utils.tensorboard.SummaryWriter(os.path.join('./logs', self.name))
 
         print("Start fitting for", self.max_iteration - self._iteration,
               "iterations / ", self.max_epoch - self._epoch, "epochs")
@@ -233,7 +175,6 @@ class DefaultTrainer:
             self.optimizer.zero_grad()
 
             prediction = self.model(x)
-            # print(prediction.min(), prediction.max())
             if self._iteration % self.log_image_interval == 0:
                 prediction.retain_grad()
             loss = self.loss(prediction, y)
@@ -242,9 +183,10 @@ class DefaultTrainer:
             self.optimizer.step()
 
             lr = [pm['lr'] for pm in self.optimizer.param_groups][0]
-            self.log_tensorboard_train(loss, lr,
-                                       x, y, prediction,
-                                       log_gradients=True)
+            if self.logger is not None:
+                self.logger.log_train(loss, lr,
+                                      x, y, prediction,
+                                      log_gradients=True)
 
             self._iteration += 1
             n_iter += 1
@@ -274,8 +216,9 @@ class DefaultTrainer:
             self.scaler.update()
 
             lr = [pm['lr'] for pm in self.optimizer.param_groups][0]
-            self.log_tensorboard_train(loss, lr,
-                                       x, y, prediction)
+            if self.logger is not None:
+                self.logger.log_train(loss, lr,
+                                      x, y, prediction)
 
             self._iteration += 1
             n_iter += 1
@@ -301,8 +244,9 @@ class DefaultTrainer:
 
         metric /= len(self.val_loader)
         loss /= len(self.val_loader)
-        self.log_tensorboard_validation(metric, loss,
-                                        x, y, prediction)
+        if self.logger is not None:
+            self.logger.log_validation(metric, loss,
+                                       x, y, prediction)
         return metric
 
     def _validate_mixed(self):
@@ -323,6 +267,7 @@ class DefaultTrainer:
 
         metric_val /= len(self.val_loader)
         loss_val /= len(self.val_loader)
-        self.log_tensorboard_validation(metric_val, loss_val,
-                                        x, y, prediction)
+        if self.logger is not None:
+            self.logger.log_validation(metric, loss,
+                                       x, y, prediction)
         return metric_val
