+++ torch_em/transform/__init__.py
@@ -1,3 +1,3 @@
 from .augmentation import get_augmentations
-from .label import AffinityTransform
+from .label import AffinityTransform, BoundaryTransform, labels_to_binary
 from .raw import get_raw_transform
